Finalize compilation effort of HIP on Windows
It compiles! Many parts are not implemented but most of the kernels do compile.
Could not run the program since I do not posses AMD GPU.

Add custom Optional type which is more restrictive than std::optional
Add macro for CPU-side profiler annotation since when system compiled without MRAY_ENABLE_TRACY, there would be many unused
variable warnings for the empty classes.

+++ Source/Core/System.cpp
@@ -94,7 +94,7 @@ std::string GetCurrentThreadName()
         mbstate_t state = {};
         static constexpr size_t MAX_TASK_COMM_LEN = 15;
         std::string result(MAX_TASK_COMM_LEN + 1, '\0');
-        size_t inputSize = lstrlenW(desc);
+        size_t inputSize = size_t(lstrlenW(desc));
         size_t totalConv = 0;
         const wchar_t* constDesc = desc;
         wcsrtombs_s(&totalConv,
